refactor preview components
