+++ src/components/Preview/Preview.tsx
@@ -21,20 +21,14 @@ import {
 import { css, cx } from "styled-system/css";
 import { center, grid, hstack, vstack } from "styled-system/patterns";
 
+import { ImagesContext } from "../../context/ImagesContext";
+import { SettingsContext } from "../../context/SettingsContext";
+import { usePreviewData } from "../../hooks/usePreviewData";
+import { getIsSortableCardData } from "../../hooks/useSortableCard";
+import { ProgressOverlay } from "../ProgressOverlay";
+import { Link } from "../ui/link";
+import { Actions } from "./Actions";
 import { Card } from "./Card";
-import { ImageErrors } from "./components/ImageErrors";
-import { ProgressOverlay } from "./components/ProgressOverlay";
-import { Button } from "./components/ui/button";
-import { Link } from "./components/ui/link";
-import { Pagination } from "./components/ui/pagination";
-import { Tooltip } from "./components/ui/tooltip";
-import { ImagesContext } from "./context/ImagesContext";
-import { SettingsContext } from "./context/SettingsContext";
-import { useGeneratePdf } from "./hooks/useGeneratePdf";
-import { useImageLoadingProgress } from "./hooks/useImageLoadingProgress";
-import { usePreviewData } from "./hooks/usePreviewData";
-import { getIsSortableCardData } from "./hooks/useSortableCard";
-import { progressEvents } from "./utils/progress-events";
 
 const containerStyles = css.raw({
   flex: 1,
@@ -204,20 +198,11 @@ const PageDrop = ({
   );
 };
 
-export const PrintableImages = () => {
+export const Preview = () => {
   const { cssVars } = useContext(SettingsContext);
 
-  const {
-    images,
-    onClear,
-    isRendering,
-    setIsRendering,
-    onClearErrors,
-    imagesWithError,
-    onReorder,
-  } = useContext(ImagesContext);
-
-  const isLoadingImages = useImageLoadingProgress();
+  const { images, isRendering, imagesWithError, onReorder } =
+    useContext(ImagesContext);
 
   const contentRef = useRef<HTMLDivElement>(null);
 
@@ -239,18 +224,6 @@ export const PrintableImages = () => {
     () => currentPage === imageMatrix.length,
     [currentPage, imageMatrix.length],
   );
-
-  const generatePdf = useGeneratePdf(contentRef);
-
-  const handleSave = () => {
-    setIsRendering(true);
-    console.time("save");
-    progressEvents.emit("progress", {
-      progress: 0,
-      phase: "Initializing",
-    });
-    generatePdf();
-  };
 
   const [dragOverlayOffset, setDragOverlayOffset] = useState<{
     x: number;
@@ -411,99 +384,12 @@ export const PrintableImages = () => {
               gap: "3",
             })}
           >
-            <div
-              className={vstack({
-                gap: "2",
-                width: "max(var(--page-width), 8.5in)",
-                minWidth: "max",
-                maxWidth: "full",
-                alignItems: "stretch",
-                position: "sticky",
-                left: "0",
-              })}
-            >
-              <div
-                className={hstack({
-                  gap: "2",
-                  justifyContent: "space-between",
-                  alignItems: "flex-end",
-                })}
-              >
-                <div className={hstack({ gap: "2" })}>
-                  <Button
-                    colorPalette="gray"
-                    disabled={isRendering}
-                    onClick={() => onClear()}
-                  >
-                    Remove all cards
-                  </Button>
-                  <Tooltip.Root
-                    disabled={
-                      !isRendering && !isLoadingImages && isReferenceCardLoaded
-                    }
-                    positioning={{
-                      placement: "top",
-                    }}
-                  >
-                    <Tooltip.Trigger asChild>
-                      <Button
-                        disabled={
-                          isRendering ||
-                          isLoadingImages ||
-                          !isReferenceCardLoaded
-                        }
-                        onClick={() => handleSave()}
-                      >
-                        Save
-                      </Button>
-                    </Tooltip.Trigger>
-                    <Tooltip.Positioner>
-                      <Tooltip.Arrow>
-                        <Tooltip.ArrowTip />
-                      </Tooltip.Arrow>
-                      <Tooltip.Content>
-                        {isRendering
-                          ? "Generating PDF..."
-                          : isLoadingImages
-                            ? "Downloading images..."
-                            : !isReferenceCardLoaded
-                              ? "Loading images..."
-                              : ""}
-                      </Tooltip.Content>
-                    </Tooltip.Positioner>
-                  </Tooltip.Root>
-                </div>
-                <div
-                  className={vstack({
-                    alignItems: "center",
-                    gap: "2",
-                    visibility: imageMatrix.length > 1 ? "visible" : "hidden",
-                  })}
-                >
-                  <span className={css({ fontSize: "xs", color: "fg.muted" })}>
-                    Showing {currentPage * cardsPerPage - cardsPerPage + 1} -{" "}
-                    {Math.min(
-                      currentPage * cardsPerPage,
-                      imageMatrix.length * cardsPerPage,
-                      images.length,
-                    )}{" "}
-                    of{" "}
-                    {Math.min(imageMatrix.length * cardsPerPage, images.length)}
-                  </span>
-                  <Pagination
-                    count={imageMatrix.length * cardsPerPage}
-                    page={currentPage}
-                    pageSize={cardsPerPage}
-                    siblingCount={1}
-                    onPageChange={({ page }) => changePage(page)}
-                  />
-                </div>
-              </div>
-              <ImageErrors
-                onDismiss={onClearErrors}
-                imagesWithError={imagesWithError}
-              />
-            </div>
+            <Actions
+              contentRef={contentRef}
+              isReferenceCardLoaded={isReferenceCardLoaded}
+              currentPage={currentPage}
+              changePage={changePage}
+            />
             <div
               ref={contentRef}
               style={
